Fixed remaining race condition in lock
If the mtime of two lock files is the same (possible), we need to have a
well-defined resolution of which lock wins.  So, we sort by the
filename, and thus the UUID.  It may not guarantee that the "first"
requestor wins (the mtime could be identical due to buffering, even if
there is a well-defined order to the locks otherwise; we were hitting
this in the test), but everyone will at least agree on the winner.